ISAICP-3166: Migrate the documentation field in 'solution' and 'release'.
